+++ backend/turip-app/src/main/java/turip/infrastructure/client/PlaceSearchClient.java
@@ -1,6 +1,7 @@
-package turip.place.domain;
+package turip.infrastructure.client;
 
 import turip.place.controller.dto.response.PlaceSearchResponse;
+import turip.place.domain.PlaceSearchType;
 
 public interface PlaceSearchClient {
     PlaceSearchResponse search(String query);
