+++ backend/turip-app/src/test/java/turip/infrastructure/client/KakaoPlaceSearchClientTest.java
@@ -1,4 +1,4 @@
-package turip.place.infrastructure.client;
+package turip.infrastructure.client;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
@@ -11,13 +11,16 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.test.autoconfigure.web.client.RestClientTest;
+import org.springframework.context.annotation.Import;
 import org.springframework.http.MediaType;
 import org.springframework.test.context.ActiveProfiles;
 import org.springframework.test.web.client.MockRestServiceServer;
+import turip.common.configuration.RestClientConfiguration;
 import turip.place.controller.dto.response.PlaceSearchResponse;
 import turip.place.domain.PlaceSearchProvider;
 
 @ActiveProfiles("test")
+@Import(RestClientConfiguration.class)
 @RestClientTest(KakaoPlaceSearchClient.class)
 class KakaoPlaceSearchClientTest {
